sbull: Use blk_fetch_request instead blk_peek_request to dequeue requests
The block layer request API changes make a requirement for block drivers
to dequeue requests before signal its completition.

Commit 1dfd4ab138fc9f9cad869e6110022c7cfd5544d1 changed sbull to use
the new block layer API. But the patch only renamed the functions while
keeping the old driver semantics. It used blk_peek_request() that does the
same that elv_next_request() did.

Remove the request from the queue before is not only a recomendation
but a true requirement. Attempting to complete a request that remains
in the queue make the kernel oops. So the last commit introduced a Bug.

This patch uses blk_fetch_request() instead of blk_peek_request().
blk_fetch_request() not only returns the request in the head of the
queue but also removes the request from the queue so a latter
signaling of the request completition doesn't hang the system.