README: update project goal according to reality
The initial goal of this project was to update the LDD3
example drivers to recent kernels so they can be used by
students to learn about driver development under Linux.

The idea was to not only update the drivers but also add
new drivers that show newer kernel facilities that did
not exist (or was not common) by the time the LDD3 book
was written.

Unfortunately I haven't had time to work on this for
years so it's better to just try to keep the current
drivers building and working.

Also, only maintain the master branch that should track
Linus master tree. Anyone that is interested on kernel
hacking should develop against the latest master tree
anyways. If someone needs to use an older kernel version
it shouldn't be hard to figure out what commit id could
be used for any specific kernel version.

Signed-off-by: Javier Martinez Canillas <javier@dowhile0.org>