Fix static BUS_ATTR_RO(version); first arg is now const
Tested on kernel v6.6.88, build was failing with:

```
/root/lkmc/submodules/ldd3/lddbus/lddbus.c:85:20: error: initialization of 'ssize_t (*)(const struct bus_type *, char *)' {aka 'long int (*)(const struct
 bus_type *, char *)'} from incompatible pointer type 'ssize_t (*)(struct bus_type *, char *)' {aka 'long int (*)(struct bus_type *, char *)'} [-Werror=i
ncompatible-pointer-types]
   85 | static BUS_ATTR_RO(version);
      |                    ^~~~~~~
```

This commit fixes that and that file now builds. The rest of the build
continues to fail on other errors that follow as before.

I've also tested the build on kernel v5.10.236, the latest patch on 5.10
which the README says should work, and it still builds.